Add utils & profile manager to @threefold/common & update playground project
