+++ packages/common/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@threefold/common",
   "version": "2.1.0-rc8",
-  "private": false,
+  "private": true,
   "scripts": {
     "dev": "vite",
     "build": "run-p type-check build-only",
