+++ packages/common/src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as TfChainWalletConnector } from "./profile_manager.vue";
+export * from "./dynamic_tabs.vue";
