+++ packages/common/src/components/qrcode_generator.vue
