+++ packages/common/src/global-components.d.ts
@@ -1,3 +1,5 @@
+import type { NetworkEnv } from "@threefold/grid_client";
+
 import CopyInputWrapper from "../components/copy_input_wrapper.vue";
 import DTabs from "../components/dynamic_tabs.vue";
 import FormValidator from "../components/form_validator.vue";
@@ -18,5 +20,22 @@ export declare module "@vue/runtime-core" {
 
   export interface ComponentCustomProperties {
     validators: ValidatorsType;
+  }
+}
+
+declare global {
+  interface Window {
+    env: {
+      NETWORK: NetworkEnv;
+      GRAPHQL_URL: string;
+      GRIDPROXY_URL: string;
+      SUBSTRATE_URL: string;
+      ACTIVATION_SERVICE_URL: string;
+      RELAY_DOMAIN: string;
+      BRIDGE_TFT_ADDRESS: string;
+      STELLAR_NETWORK: string;
+      STELLAR_HORIZON_URL: string;
+      TFT_ASSET_ISSUER: string;
+    };
   }
 }
