+++ packages/common/src/types/index.ts
@@ -1,0 +1,6 @@
+export interface NodeGPUCardType {
+  id: string;
+  vendor: string;
+  device: string;
+  contract: number;
+}
