+++ packages/common/src/utils/index.ts
@@ -1,5 +1,6 @@
 import * as validators from "./validators";
 
+export * from "./grid";
+export * from "./helpers";
 export type ValidatorsType = typeof validators;
-
 export { validators };
