+++ packages/playground/src/components/k8s_deployment_table.vue
@@ -58,10 +58,9 @@
 </template>
 
 <script lang="ts" setup>
+import { getGrid, updateGrid, useProfileManager } from "@threefold/common";
 import { onMounted, ref } from "vue";
 
-import { useProfileManager } from "../stores";
-import { getGrid, updateGrid } from "../utils/grid";
 import { loadK8s, mergeLoadedDeployments } from "../utils/load_deployment";
 
 const profileManager = useProfileManager();
