--- packages/playground/src/stores/index.ts
@@ -1,1 +1,0 @@
-export { useProfileManager } from "./profile_manager";
