+++ packages/playground/src/types/index.ts
@@ -1,8 +1,7 @@
+import type { ValidatorsType } from "@threefold/common";
 import type { VDataTable } from "vuetify/lib/labs/components";
 
 import type { INode } from "@/utils/filter_nodes";
-
-import type * as validators from "../utils/validators";
 
 export interface K8SWorker {
   name: string;
@@ -137,4 +136,4 @@ export interface SMTPServer {
   ssl: boolean;
 }
 
-export type Validators = typeof validators;
+export type Validators = ValidatorsType;
