+++ packages/playground/src/weblets/tf_node_pilot.vue
@@ -192,12 +192,13 @@ async function deploy() {
 </script>
 
 <script lang="ts">
+import { normalizeError } from "@threefold/common";
+
 import SelectFarmId from "../components/select_farm.vue";
 import SelectFarmManager from "../components/select_farm_manager.vue";
 import SelectNode from "../components/select_node.vue";
 import { deploymentListEnvironments } from "../constants";
 import type { INode } from "../utils/filter_nodes";
-import { normalizeError } from "../utils/helpers";
 
 export default {
   name: "NodePilot",
