+++ packages/playground/src/weblets/tf_taiga.vue
@@ -306,9 +306,10 @@ async function deploy(gatewayName: GatewayNode, customDomain: boolean) {
 </script>
 
 <script lang="ts">
+import { normalizeError } from "@threefold/common";
+
 import DomainName from "../components/domain_name.vue";
 import FarmGatewayManager from "../components/farm_gateway_manager.vue";
-// import Networks from "../components/networks.vue";
 import SelectFarm from "../components/select_farm.vue";
 import SelectFarmManager from "../components/select_farm_manager.vue";
 import SelectNode from "../components/select_node.vue";
@@ -316,7 +317,6 @@ import SelectSolutionFlavor from "../components/select_solution_flavor.vue";
 import SmtpServer, { createSMTPServer } from "../components/smtp_server.vue";
 import { deploymentListEnvironments } from "../constants";
 import type { INode } from "../utils/filter_nodes";
-import { normalizeError } from "../utils/helpers";
 import rootFs from "../utils/root_fs";
 
 export default {
